refactor view elements due to new scripts in system
